% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_data}
\alias{cache_data}
\title{Cache files read in by the API}
\usage{
cache_data(name, data, reset = FALSE, environment = .FAOSTATenv)
}
\arguments{
\item{name}{character. Key to which data is bound}

\item{data}{object. Data to be cached}

\item{reset}{logical. Should the cache be replaced with new data?}

\item{environment}{environment. .FAOSTATenv by default}
}
\description{
Cache files read in by the API
}
\keyword{internal}
