% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CHMT}
\alias{CHMT}
\title{This function avoids double counting of China.}
\usage{
CHMT(var, data, year = "Year")
}
\arguments{
\item{var}{The variables that require to be sanitized.}

\item{data}{The data frame which contains the data}

\item{year}{The column which correspond to the year.}
}
\description{
This function should only be used when performing aggregations.
}
\details{
We decide to use the smaller subsets in the regional level because
weighting variable may not exist for other variables for the
larger subsets.

The function only work for FAOST_CODE, if the country coding
system is not in FAOST_CODE then use the translateCountryCode
function to translate it.
}

