% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_metaTable_updated.R
\docType{data}
\name{fao_metaTable_updated}
\alias{fao_metaTable_updated}
\title{Link to cached version of fao_metaTable_updated}
\format{
A data frame with 74 rows and 12 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  \item{DatasetCode}{Each dataset with unique code}
  \item{DatasetCode}{Full Dataset name}
  \item{Topic}{Brief description of Dataset}
  \item{DatasetDescription}{Non-brief Dataset description}
  \item{DateUpdate}{Date of latest update}
  \item{FileSize}{Size of Dataset}
  \item{FileLocation}{Direct URL for file download}
  ...
}
}
\source{
\url{http://fenixservices.fao.org/faostat/static/bulkdownloads/datasets_E.xml}
}
\description{
A full table containing the latest cached datasets available via the FAOSTATs webservice
}
\keyword{datasets}
