\name{FAOcheck}
\alias{FAOcheck}
\title{This function perform some check on the data}
\usage{
  FAOcheck(var, year = "Year", data,
    type = c("overlap", "multiChina"))
}
\arguments{
  \item{var}{The variable to be checked}

  \item{year}{The column which index the time.}

  \item{data}{The data frame}

  \item{type}{The type of check}
}
\description{
  This function perform some check on the data
}

