% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAOmetaTable.R
\docType{data}
\name{FAOmetaTable}
\alias{FAOmetaTable}
\title{The search tree for FAOSTAT}
\description{
A table containing the relationship between the domain, element,
item, and area codes for downloading data from the FAOSTAT API.
}
\keyword{datasets}

