% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FAOregionProfile.R
\docType{data}
\name{FAOregionProfile}
\alias{FAOregionProfile}
\title{Regional profile}
\description{
Region profile containing the codes, names and regional
classifications of countries.
}
\keyword{datasets}

