% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAOsearch.R
\name{FAOsearch}
\alias{FAOsearch}
\title{Search FAOSTAT tables}
\usage{
FAOsearch(
  code = NULL,
  dataset = NULL,
  topic = NULL,
  latest = FALSE,
  full = FALSE
)
}
\arguments{
\item{code}{Specific 2-3 character code for dataset, listed as DatasetCode}

\item{dataset}{Name of dataset, listed as DatasetName}

\item{topic}{Dataset topic from list}

\item{latest}{Sort list by latest updates}

\item{full}{defaults False, return the full table with all columns}
}
\description{
Get full list of datasets from the FAOSTAT database with the Code, Dataset Name and Topic.
}
