% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAOSTAT-package.R
\docType{package}
\name{FAOSTAT-package}
\alias{FAOSTAT}
\alias{FAOSTAT-package}
\title{A complementary package to the FAOSTAT database and the
Statistical Yearbook of the Food and Agricultural Organization of
the United Nations.}
\description{
Download Data from the FAOSTAT Database of the Food and Agricultural Organization (FAO) of the United Nations. A list of functions to download statistics from FAOSTAT (database of the FAO \url{https://www.fao.org/faostat/}) and WDI (database of the World Bank \url{https://data.worldbank.org/}), and to perform some harmonization operations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/paulrougieux/faostatpackage}
  \item Report bugs at \url{https://gitlab.com/paulrougieux/faostatpackage/-/issues}
}

}
\keyword{package}
