% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAOSTAT.R
\name{FAOSTAT}
\alias{FAOSTAT}
\title{Access to FAOSTAT data}
\usage{
FAOSTAT(name = NULL, domainCode = "QC", elementCode = 2510,
  itemCode = NULL, yearRange = NULL, countrySet = NULL, query,
  toDataFrame = FALSE)
}
\arguments{
\item{name}{The name to be given to the variable.}

\item{domainCode}{The domain code of the variable, see details.}

\item{elementCode}{The element code of the variable, see details.}

\item{itemCode}{The item code of the variable, see details.}

\item{yearRange}{A numeric vector containing the years to be downloaded.}

\item{countrySet}{A numeric vector of FAOSTAT codes of those countries to be downloaded.}

\item{query}{The object created if using the FAOsearch function.}

\item{toDataFrame}{Logical, whether to return the results in data frames instead of data tables.}
}
\value{
A list containing the following elements
\describe{
    \item{entity}{The entity level data}
    \item{aggregates}{The aggregates provided by the FAO}
    \item{results}{The status of the download, whether success/failed}
}
}
\description{
A wrapper function to obtain multiple data sets under the FAOSTAT 
domains: "Production"; "Trade" (excluding "Detailed trade matrix");
"Food Balance"; the subdomain "Suite of Food Security Indicators"
under the domain "Food Security"; "Prices"; "Inputs" (excluding 
"Employment Indicators"); "Population"; "Investment" (excluding "Development 
Flows to Agriculture"); "Macro-Statistics"; "Agri-Environmental Indicators";
"Emissions - Agriculture"; "Emissions - Land Use"; the subdomain 
"Forestry Production and Trade" under the domain "Forestry"; "ASTI R&D Indicators".
}
\examples{

## production.dt <- 
## data.table(varName = c("Grapes_AreaHarv", 
##                       "Wine_ProdQuantity", 
##                       "CattleBuffaloes_Stocks", 
##                       "MilkProducts_Yield",
##                       "Yoghurt_ProdQuantity",
##                       "Food _GPIN",
##                       "Agriculture_NetProdValue"),
##           domainCode = c("QC", "QD", "QA", "QL", "QP", "QI", "QV"),
##           itemCode = c(560, 564, 1746, "1780>", 891, 2054, 2051),
##           elementCode = c(2312, 2510, 2111, 2413, 2510, 432, 154))
## production.lst <- with(production.dt,
##                       FAOSTAT(name = varName, domainCode = domainCode,
##                               itemCode = itemCode, elementCode = elementCode,
##                               yearRange = c(2005:2010), countrySet = c(106, 5400)))

}

