% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAOSTATsearch.R
\name{FAOSTATsearch}
\alias{FAOSTATsearch}
\title{A function to find domain, element, item, and area code for a
specific FAOSTAT query.}
\usage{
FAOSTATsearch()
}
\description{
A function to find domain, element, item, and area code for a
specific FAOSTAT query.
}

