% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAOSTATtrade.R
\name{FAOSTATtrade}
\alias{FAOSTATtrade}
\title{Access to FAOSTAT data on trade matrix}
\usage{
FAOSTATtrade(name = NULL, domainCode = "QC", elementCode = 2510,
  itemCode = NULL, yearRange = NULL, countrySet = NULL, query,
  toDataFrame = FALSE)
}
\arguments{
\item{name}{The name to be given to the variable.}

\item{domainCode}{The domain code of the variable.}

\item{elementCode}{The element code of the variable.}

\item{itemCode}{The item code of the variable.}

\item{yearRange}{A numeric vector containing the years to be downloaded.}

\item{countrySet}{A numeric vector of FAOSTAT codes of those countries to be downloaded.}

\item{query}{The object created if using the FAOsearch function.}

\item{toDataFrame}{Logical, whether to return the results in data frames instead of data tables.}
}
\value{
A list containing the following elements
\describe{
    \item{entity}{The entity level data}
    \item{aggregates}{The aggregates provided by the FAO}
    \item{results}{The status of the download, whether success/failed}
}
}
\description{
A wrapper function to obtain multiple
data sets under the FAOSTAT domains "Detailed trade matrix" and
"Forestry Trade Flows".
}
\examples{
## The default option is production in quantity
## wineExp.lst <- FAOSTATtrade(domainCode = "TM", elementCode = 2910, itemCode = 564, 
##                             yearRange = 2000:2002, countrySet = c(68,106))
## sawnwoodImp.lst <- FAOSTATtrade(domainCode = "FT", elementCode = 2610, itemCode = 1632, 
##                                 yearRange = 2013:2014, countrySet = c(21,44))
}

