% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{get_fao}
\alias{get_fao}
\title{REST functions to connect to FAO}
\usage{
get_fao(
  endpoint,
  params = list(),
  language = c("en", "fr", "es", "NA"),
  base_url = "https://faostatservices.fao.org/api/v1"
)
}
\arguments{
\item{endpoint}{character. Endpoint for FAO API}

\item{params}{list of query parameters}

\item{language}{character. One of the UN languages supported by FAOSTAT}

\item{base_url}{character. Base URL for FAOSTAT API}
}
\description{
REST functions to connect to FAO
}
\keyword{internal}
