% Generated by roxygen2 (4.0.1): do not edit by hand
\name{overlap}
\alias{overlap}
\title{This function checks whether there are overlapping between the
transitional countries.}
\usage{
overlap(old, new, var, year = "Year", data)
}
\arguments{
\item{old}{The FAOST_CODE of the old countries}

\item{new}{The FAOST_CODE of the new countries}

\item{var}{The variable to be checked}

\item{year}{The column which index the time.}

\item{data}{The data frame}
}
\description{
This function checks whether there are overlapping between the
transitional countries.
}

