\name{printLab}
\alias{printLab}
\title{Print labels}
\usage{
  printLab(label, span = FALSE, width = getOption("width"))
}
\arguments{
  \item{label}{The label to be printed}

  \item{span}{Whether the dash should span the whole width
  of the screen(80 characters)}

  \item{width}{The width of the screen.}
}
\value{
  The formatted print
}
\description{
  A function to print standardised formatted labels without
  having messy codes in the functions.
}

