% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset_dimension.R
\name{read_dataset_dimension}
\alias{read_dataset_dimension}
\alias{read_dimension_metadata}
\title{Examine dimensions of a dataset}
\usage{
read_dataset_dimension(dataset_code)

read_dimension_metadata(dataset_code, dimension_code)
}
\arguments{
\item{dataset_code}{character. Dataset as obtained from the code column of
\link{search_dataset}}

\item{dimension_code}{character. Dimensions as obtained from \code{read_dataset_dimensions}}
}
\description{
Lists the dimensions of a dataset including ids and labels. These can be used
to query dataset dimension names and the codes therein. They can also be used
to access groups, flags, units and the glossary
}
