% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sparsity}
\alias{sparsity}
\title{A function to calculate the sparsity of the variable}
\usage{
sparsity(year, var, data)
}
\arguments{
\item{year}{The column which index the time}

\item{var}{The variable to calculate sparsity}

\item{data}{The data frame.}
}
\description{
When two or more variables are merged together, it is very likely
that spurious sparsity is created. This is due to the fact that
the years between the variables are not identicle and is
particularly true for data stored as data.frame with country and
year as column indexes.
}
\details{
This function computes the sparsity by taking the data which are
observed between the first and the last observation.

For example, if variable A is observed since 1960 while variable B
is collected since 1990. When they are merged together, NA's will
be created for variable prior to 1990 and these should not be used
to calculate the sparsity of the data.
}

