% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sparsityHeatMap}
\alias{sparsityHeatMap}
\title{Function for examine sparsity}
\usage{
sparsityHeatMap(data, country, year, var, group, ncol)
}
\arguments{
\item{data}{The data frame containing the data.}

\item{country}{The column name containing the country name.}

\item{year}{The columne name containing the time information.}

\item{var}{The variable to be examined.}

\item{group}{The colomun name which the data should be grouped, if
missing quantile of the variable is used.}

\item{ncol}{The number of column for the resulting plot}
}
\description{
The function plots the availability of a specific variable over time
by country and are grouped by the average value of the indicator.
}

