% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_countrycode.R
\name{translate_countrycode}
\alias{translate_countrycode}
\alias{translateCountryCode}
\title{A function to translate between different country coding systems}
\usage{
translate_countrycode(
  data,
  from = c("FAO", "M49", "ISO2", "ISO3"),
  to = c("M49", "FAO", "ISO2", "ISO3", "name"),
  oldCode,
  reset_cache = FALSE
)

translateCountryCode(
  data,
  from = c("FAO", "M49", "ISO2", "ISO3"),
  to = c("M49", "FAO", "ISO2", "ISO3", "name"),
  oldCode,
  reset_cache = FALSE
)
}
\arguments{
\item{data}{The data frame}

\item{from}{The name of the old coding system}

\item{to}{The name of the new coding system}

\item{oldCode}{The column name of the old country coding scheme}

\item{reset_cache}{logical. Whether to pull data from FAOSTAT directly instead of caching}
}
\description{
The function translate any country code scheme to another if both
are in the FAOcountryProfile. These wil
}
