% Generated by roxygen2 (4.0.1): do not edit by hand
\name{translateUnit}
\alias{translateUnit}
\title{Function to translate multipliers}
\usage{
translateUnit(vec)
}
\arguments{
\item{vec}{The vector containing name or number to be translated}
}
\description{
This function translates number to character name or vice versa
}
\examples{
## Create numeric vector
myUnit = c(1000, 1e6, 1000, 1e9, 1e9, 1e12)

## Translate numeric to character
myUnit2 = translateUnit(myUnit)
myUnit2

## Now translate back
translateUnit(myUnit2)
}

