% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{check_cache}
\alias{check_cache}
\title{Cache files read in by the API}
\usage{
check_cache(name, environment = .FAOSTATenv)
}
\arguments{
\item{name}{character. Key to which data is bound}

\item{environment}{environment. .FAOSTATenv by default}

\item{data}{object. Data to be cached}

\item{reset}{logical. Should the cache be replaced with new data?}
}
\description{
Cache files read in by the API
}
\keyword{internal}
