% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chConstruct.R
\name{chConstruct}
\alias{chConstruct}
\title{Construct year to year change}
\usage{
chConstruct(data, origVar, country = "FAOST_CODE", year = "Year",
  newVarName = NA, n = 1)
}
\arguments{
\item{data}{The data frame containing the data}

\item{origVar}{The variable in which the year to year change is to
be calculated}

\item{country}{The column representing the index of country.}

\item{year}{The column represing the index of year.}

\item{newVarName}{The name assigned to the new variable, if
missing then .CH will be appended.}

\item{n}{The period for the change rate to be calculated.}
}
\value{
A data frame containing the computed year to year change
rate.
}
\description{
A function for constructing year to year change
}

