% Generated by roxygen2 (4.0.1): do not edit by hand
\name{FAOsearch}
\alias{FAOsearch}
\title{A function to find the domain, element and item code for a
specific FAOSTAT query.}
\usage{
FAOsearch()
}
\description{
A function to find the domain, element and item code for a
specific FAOSTAT query.
}

