% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAOSTAT-package.R
\docType{package}
\name{FAOSTAT-package}
\alias{FAOSTAT-package}
\alias{.FAOSTATenv}
\title{A complementary package to the FAOSTAT database and the
Statistical Yearbook of the Food and Agricultural Organization of
the United Nations.}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.FAOSTATenv
}
\description{
A complementary package to the FAOSTAT database and the
Statistical Yearbook of the Food and Agricultural Organization of
the United Nations.
}
\author{
Michael. C. J. Kao \email{michael.kao@fao.org}
}
\keyword{package}
