% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAOSTAT-package.R
\docType{package}
\name{FAOSTAT-package}
\alias{FAOSTAT-package}
\title{Search and extract data from FAOSTAT database of the Food and 
Agricultural Organization of the United Nations.}
\description{
Search and extract data from FAOSTAT database of the Food and 
Agricultural Organization of the United Nations.
}
\author{
Michael. C. J. Kao \email{michael.kao@fao.org}
}
\keyword{package}

