% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFAO}
\alias{getFAO}
\title{Access to FAO FAOSTAT API.}
\usage{
getFAO(name = NULL, domainCode = "RL", elementCode = 5110,
  itemCode = 6621, query, printURL = FALSE, useCHMT = TRUE,
  outputFormat = "wide", returnNames = FALSE, returnFlags = FALSE,
  yearRange = NULL, countrySet = NULL)
}
\arguments{
\item{name}{The name to be given to the variable.}

\item{domainCode}{The domain of the data.}

\item{elementCode}{The code of the element.}

\item{itemCode}{The code of the specific item.}

\item{yearRange}{A numeric vector containing the years to be downloaded.}

\item{countrySet}{The FAOSTAT codes of those countries to be downloaded.}

\item{query}{The object created if using the FAOsearch function.}

\item{printURL}{Whether the url link for the data should be printed.}

\item{useCHMT}{logical, whether the CHMT function should be
applied to avoid double counting of China.}

\item{outputFormat}{The format of the data, can be 'long' or 'wide'.}

\item{returnNames}{Logical, should the area, the element and the
item names be reported?.}

\item{returnFlags,}{Logical, whether the flags should be
returned. Only work with outputFormat long.}
}
\value{
Outputs a data frame containing the specified data.
}
\description{
A function to access FAOSTAT data through the FAOSTAT API.
}
\details{
Need to account for multiple itemCode, currently only support one
single variable.
}
\seealso{
\code{\link{getWDI}}, \code{\link{getWDItoSYB}},
\code{\link{getFAOtoSYB}}, \code{\link{FAOsearch}}
}

