% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{This function checks whether there are overlapping between the
transitional countries.}
\usage{
overlap(old, new, var, year = "Year", data, take)
}
\arguments{
\item{old}{The FAOST_CODE of the old countries}

\item{new}{The FAOST_CODE of the new countries}

\item{var}{The variable to be checked}

\item{year}{The column which index the time.}

\item{data}{The data frame}

\item{take}{The type of check/replacement to be done.}
}
\description{
This function checks whether there are overlapping between the
transitional countries.
}
