% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_datasets.R
\name{search_datasets}
\alias{search_datasets}
\alias{FAOsearch}
\title{Search FAOSTAT domains, datasets, elements, indicators, and items}
\usage{
search_datasets(code, dataset, latest = TRUE)

FAOsearch(code, dataset, latest = TRUE)
}
\arguments{
\item{code}{character. Codes of desired datasets, listed as DatasetCode}

\item{dataset}{character. Name of the dataset (or part of the name), listed as label in the output data frame}

\item{latest}{logical. Sort list by latest updates}
}
\value{
A data.frame with the columns:
 code, label, date_update, note_update, release_current, state_current, year_current, release_next, state_next, year_next
}
\description{
Get full list of datasets from the FAOSTAT database with the Code, dataset name and updates.
}
\examples{
\dontrun{
# Find information about all datasets
fao_metadata <- search_datasets()    
# Find information about the forestry dataset
search_datasets(code="FO")
# Find information about datasets whose titles contain the word "Flows"
search_datasets(dataset="Flows")
}

}
